backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.9
firebase-admin==6.5.0
qrcode[pil]==7.4.2
Pillow==10.4.0
PyJWT==2.9.0
httpx==0.27.2
pytest==8.3.3
email-validator==2.2.0
